requirements.txt
requests>=2.31.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
Pillow>=10.0.0
pytesseract>=0.3.10
beautifulsoup4>=4.12.0
python-telegram-bot[job-queue]>=21.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
